+++ tests/phpunit/wpmatomo/test-updater.php
@@ -83,6 +83,8 @@ class UpdaterTest extends MatomoAnalytics_TestCase {
 
 	public function test_update_does_not_fail() {
 		$this->updater->update();
+		// for the phpunit warning
+		$this->assertTrue( true );
 	}
 
 }