+++ tests/phpunit/wpmatomo/user/test-sync.php
@@ -77,6 +77,10 @@ class UserSyncTest extends MatomoAnalytics_TestCase {
 	 * @group ms-required
 	 */
 	public function test_sync_all_passes_correct_values_to_sync_site_when_there_are_multiple_blogs() {
+		if ( ! is_multisite() ) {
+			$this->markTestSkipped( 'Not multisite.' );
+			return;
+		}
 		$blogid1   = self::factory()->blog->create( array( 'domain' => 'foobar.com' ) );
 		$blogid2   = self::factory()->blog->create( array( 'domain' => 'foobar.baz' ) );
 		$site_sync = new \WpMatomo\Site\Sync( new Settings() );